Allow jwst v1.8.x
