+++ exotic_miri/stage_0/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "IntegrationTimesStep",
+]
+
+from exotic_miri.stage_0.get_integration_times import IntegrationTimesStep
