+++ exotic_miri/stage_2/__init__.py
@@ -3,7 +3,6 @@ __all__ = [
     "ReadNoiseStep",
     "FlatFieldStep",
     "WavelengthMapStep",
-    "IntegrationTimesStep",
     "StitchChunksStep",
     "InspectDQFlagsStep",
     "CleanOutliersStep",
@@ -17,7 +16,6 @@ from exotic_miri.stage_2.get_gain import GainStep
 from exotic_miri.stage_2.get_readnoise import ReadNoiseStep
 from exotic_miri.stage_2.get_flat_field import FlatFieldStep
 from exotic_miri.stage_2.get_wavelength_map import WavelengthMapStep
-from exotic_miri.stage_2.get_integration_times import IntegrationTimesStep
 from exotic_miri.stage_2.stitch_chunks import StitchChunksStep
 from exotic_miri.stage_2.inspect_dq_flags import InspectDQFlagsStep
 from exotic_miri.stage_2.clean_outliers import CleanOutliersStep
